Refactor specs and benchmark in preparation to remove default args from `StackRecorder`
I plan to remove the default values for the `StackRecorder`
`cpu_time_enabled:` and `alloc_samples_enabled:` args.

Thus, this refactors all tests and one benchmark where the
`StackRecorder` was being manually created to make sure it provides its
own values for those arguments.

Why remove the default values? I want to force callers in production
code to be explicit about their decisions on the profile types in use,
as this is important to get right so the profiling backend can
distinguish which types of profiles are enabled.

Thus, I want to avoid accidentally enabling or disabling some profile
type because that was the default on the constructor.