[PROF-7251] Exclude disabled profiling sample value types from output
**What does this PR do?**:

Up until now, the profiler always included all sample value types in
the resulting profile, even if some value is was expected to always be
zero.

(Why would a value always be zero? The `cpu-time` value is only
available on Linux, and would always be zero on macOS; and we plan
to have a way to disable allocation profiling, and thus
`alloc-samples` will also be zero.)

This PR:

* Abstracts away from all components other than the `StackRecorder`
  which sample value types are enabled, by introducing a new
  `sample_values` struct that every other component can use instead
  of having to deal with the old array and other implementation
  details.

* Extends the `StackRecorder` to allow `cpu-time` and
  `alloc-samples` to be disabled

* Updates the profiler component creation to configure `StackRecorder`
  correctly.

**Motivation**:

Having the extra sample value types always being included as we did
so far is mostly harmless.

But, it does make one thing difficult -- detection of which profile
types are actually enabled in an application, in general.

Thus, the main reason for doing this change is so that the Datadog
profiling backend can easily detect which profile types are enabled,
and which aren't, so it can display better guidance and error
messages to customers.

**Additional Notes**:

N/A

**How to test the change?**:

The change includes test coverage.

To manually validate this, you can download a pprof profile via the
Datadog UX and confirm that only the enabled sample value types are
available in it, for instance by running
`go tool pprof -raw rubyprofile.pprof` and observing the columns
shown after the `Samples:` header -- disabled value types will not
show up there.