Reuse array for listing threads instead of creating one every time
**What does this PR do?**:

This PR changes the `ddtrace_thread_list` method to, instead of
creating a new Ruby array to contain its result, receive a
pre-existing array for it to use.

Then, it changes the `CpuAndWallTime` collector to always reuse the
same array for this list.

**Motivation**:

The thread list array was the one single Ruby object allocation that
the profiler always did on every sample. Now that we're exposing
object counts and experimenting with allocation profiling, this
regular allocation was really annoying to deal with when looking
at data and writing tests.

By reusing the same object, this problem goes away, and even some
specs can be simplified.

**Additional Notes**:

N/A

**How to test the change?**:

This change is covered by the existing unit tests.