Tweak spec to make sure there are no unexpected memory allocations
When I tweaked this test to tighten the number of expected allocations
I was testing on Ruby 2.7.

It turns out that on Ruby 3 we would still observe one or two extra
allocations (Ruby warming up some caches). I now tweaked the test to
pre-create all needed things, so that we can assert on the exact
number of expected allocations on Ruby 3 as well.