Fix small inconsistencies in ddtrace logging
**What does this PR do?**

While reviewing another PR, I noticed that we're somewhat inconsistent
when we prefix logging messages:

* In a few cases we used `[DDTRACE]` instead of `[ddtrace]`
* In a few cases we prefixed the `[ddtrace]` with a space, others
  we didn't

This PR tweaks all messages to use the same `[ddtrace]` prefix and
to not use a space before it.

**Motivation**:

Improve consistency of user-visible messages.

**Additional Notes**:

N/A

**How to test the change?**:

Some of the messages do have test coverage; for others, we'll need
to rely on the PR review to make sure the change is correct.