[PROF-7307] Add support for allocation samples to `ThreadContext`
**What does this PR do?**:

This PR adds the `thread_context_collector_sample_allocation` API to
the `ThreadContext` collector.

This will be used to support allocation profiling (but right now
it's not being called automatically when allocations happen, yet).

**Motivation**:

This is an incremental step towards supporting allocation profiling.

I'm still working on the scaling/weighting strategy, and may
end up adding that logic directly on the `ThreadContext`, haven't
decided yet.

**Additional Notes**:

N/A

**How to test the change?**:

Change includes test coverage. End-to-end testing will come later
once this is wired up in the `CpuAndWallTimeWorker`.