Fix "NameError: uninitialized constant Datadog::Core::Vendor::IPAddr::AddressFamilyError"
**What does this PR do?**:

This PR fixes a bug in the recently-introduced `IPAddr` module.

During review of #2665 the `IPAddr` module was introduced, and while
experimenting with the steep type checker, I noticed there was a
collision between our `IPAddr` module and the `IPAddr` module where
we were looking for execptions.

I'm not sure if this would happen in production, since we always
created an `IPAddr` instance with an IP, never a specific `family`,
but let's fix this anyway.

**Motivation**:

Less bugs!

**Additional Notes**:

(N/A)

**How to test the change?**:

This module is a backport of `IPAddr` to older Rubies. Previously,
we only tested it indirectly.

I added a minimal test to validate this change.