Rename conflicting send_payload method
The transport infrastructure is mostly generic, except for a few choke
points. This addresses one of them.

The `send_payload` method, while defined in a trace-specific module, is
included at the class level.

When attempting to reuse transport for other use cases, the method
conflicts with other modules extending the generic client class.

Luckily its use is very limited and self-contained to a trace transport
specific scope. By renaming it additional modules are able to be
included and can follow a consistent pattern without conflict.