Fix profiling/validate_benchmarks_spec failing with benchmark-ips 2.12
**What does this PR do?**:

This PR fixes the `spec/datadog/profiling/validate_benchmarks_spec.rb`
failing with benchmark-ips 2.12 with
`uninitialized constant Benchmark::IPS::NoopSuite (NameError)`.

The specific issue is caused by a refactor introduced in
https://github.com/evanphx/benchmark-ips/pull/111
which removed the `NoopSuite` class.

Previously, we were subclassing it to implement our own custom
"suite" which allowed reporting of benchmark-ips results via dogstatsd.

The only usefulness of this class was providing no-op implementations
for all methods that are required by benchmark-ips, so I've now just
directly added them to our class.

**Motivation**:

Fix broken CI! :)

**Additional Notes**:

This is technically a breaking change upstream :/

**How to test the change?**:

Validate that CI is back to green.