Store power in joules, and use prototype libdatadog feature to convert it to watts
It doesn't make sense to store individual samples in watts, as watts
already include a time dimension, so it doesn't make sense to sum them
across multiple samples.

Instead, we should store them in joules. But, it's really nice to be
able to see watts, so I've hacked up a libdatadog branch
(<https://github.com/datadog/libdatadog/tree/ivoanjo/prototype-joules-to-watts>)
that reads the joules and converts them to watts at serialization times.

The way it works is that the value types for both still need to be
defined, but the watts values don't need to be set -- libdatadog will
set them automatically.