Remove disabled JRuby 9.2.8.0 testing setup
**What does this PR do?**:

In #2572 we changed our CI setup to test only with JRuby 9.2.21.0,
disabling JRuby 9.2.8.0. Prior to that change, we tested with both
versions.

At the time we decided to only disable the JRuby 9.2.8.0 testing,
commenting it out.

This PR actually removes all references to testing with JRuby 9.2.8.0.

As a reminder, we recommend that all our customers use at least
JRuby 9.2.21.0, and also recommend evaluating the move to JRuby 9.3,
as the JRuby developers have marked JRuby 9.2 as end-of-life:
<https://twitter.com/jruby/status/1611063274459566106>.

**Motivation**:

Avoid having leftover setup that's not in use.

**Additional Notes**:

N/A

**How to test the change?**:

Validate that CI is still green on all configurations.