Avoid triggering allocation sampling during sampling
In rare cases (such as when throwing exceptions), the profiler may need
to allocate Ruby objects during sampling.

This object allocation will trigger the `object_allocation_tracepoint`
but we don't want to try to sample in the middle of sampling.

To avoid any issues altogether, I've introduced a `during_sample` flag
that can be used to detect this and avoid trying to do sampling
during sampling.

Note that the functions modified all run while holding on to the
global VM lock, which is why this flag does not need any special
concerns re: concurrency.

I've also chosen to keep a counter for how often this happens that
may come in handy when debugging any issues around this (and should
be cheap enough to keep).