Fix Rails < 3 conditional check
Previously, this was evaluating to "false", which is not "nil", so railtie_supported? was returning true. Railties are not supported on Rails versions prior to 3 (which is what this check was trying to accomplish), so I've removed the explicit .nil? check and replaced it with a boolean coercion (which will return false for both "nil" and "false", rather then returning false for "nil" and true for "false")