Decouple profiling component initialization from core
**What does this PR do?**:

Similar to the approach taken by the AppSec component, this PR changes
the Profiling component initialization to be self method on
`Datadog::Profiling::Component` that is completely separate from the
core components.

This also included moving over the specs for this component to a
separate file.

**Motivation**:

Finish the migration started in #2561 when the Profiling component
was moved away from the core.

**Additional Notes**:

I've also sprinkled a few documentation improvements throughout.

**How to test the change?**:

This change is covered by the included and existing specs.