[PROF-6555] Support disabling endpoint names collection in new profiler
**What does this PR do?**:

This PR updates the new CPU Profiling 2.0 profiler to honor the
existing `profiling.advanced.endpoint.collection.enabled` setting.

This setting defaults to enabled. When disabled, the profiler will omit
endpoint names in the collected data.

**Motivation**:

Not supporting this was an oversight; the new profiler (still in beta)
should honor this setting.

**Additional Notes**:

N/A

**How to test the change?**:

Change includes test coverage.