Remove check for cpu-time profiling being supported
This check is a relic of older times when the profiling native extension
did not exist.

Nowadays, the checks for JRuby and Windows and non-linux platforms have
moved to the `Datadog::Profiling.unsupported_reason`, and they will
come from `native_library_compilation_skipped?` (via the
`NativeExtensionHelpers` module).

Thus, on those platforms, the profiling code flow doesn't even try to
run the `Setup` task (so the warning would not be shown).

For the macOS setup, we're working on moving the indication that
macOS has no cpu-time to the Datadog UX itself. But I think it's OK
to drop this now, because running on macOS is anyway something that
involves manual work (such as compiling your own libdatadog), so
it's not something we currently expect customers to go through.