Add type signatures for a bunch of profiling classes
**What does this PR do?**:

This PR adds type signatures and enables type checking for a number
of profiling classes (see `Steepfile` for list of files that are no
longer ignored).

**Motivation**:

This was discussed/requested in #2697.

**Additional Notes**:

I wish the steep errors were a bit more user-friendly.

The `StackRecorder#clear` method actually should not exist anymore,
and would actually break because it called `_native_clear` which was
deleted in #2367. (Ooops)

Other than that, there's a few minor changes to code to make code more
type-checkable, but nothing of notice.

**How to test the change?**:

Validate that CI is still green and typechecking passes.