Fix (hopefully for good) flaky profiling "sampling while VM is idle" spec
**What does this PR do?**:

This PR attempts to fix a flaky profiling spec that was last seen in
<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/9405/workflows/89cdfecc-0767-4fc9-85ab-1caa0abda773/jobs/349752>

This spec has some inherent flakiness, as it runs the profiler for a
short while and then asserts on some of the counters. Since all this
is asynchronous, values could be off.

In fact, this is not the first time this spec was changed to try to
resolve flakiness:
* e194ec68dc5841d7dae68cc53c1d81e34ac9cea3
* #2539
* 7d22e545c2a1adc4856faf1956c1b5bab1338396

The last time this caused flakiness, the source was the dynamic
sampling rate component. If for some reason the CI machine is having
a slow day, sampling takes longer, and then the dynamic sampling
rate component kicks in and delays samples (this is by design!).

This has happened again today; so, as I promised on
7d22e545c2a1adc4856faf1956c1b5bab1338396, I went in an added a way
to disable the dynamic sampling rate component so it does not
impact this spec.

**Motivation**:

Get us to zero flaky specs.

**Additional Notes**:

N/A

**How to test the change?**:

Change includes test coverage.