Spawn remote configuration lazily on barrier check
This commit hides worker spawning behind a barrier declaration, which
better reflects the intent.

By spawning remote configuration on first request we ensure that the
worker is only spawned in the proper process. This makes the worker
impervious to thread forking issues (e.g should it be spawned in the
master process of a forking server like unicorn or puma)

While this makes it look like the first request is not immediately
consistent with remote configuration, even if we move it earlier the
first remote configuration responses may be empty for a while (due to
e.g API rate limiting, agent caching, backend delays in face of a new
client id...)

Therefore starting the worker earlier would not help in guaranteeing the
first requests are operating with settings in sync with remote
configuration, and a barrier would still need to be put on the first
request since remote configuration update is done on a separate thread.

It should be possible to make `barrier` blocking until the first
non-empty remote configuration update is applied, but the case of
legitimate forever-empty remote configuration must be considered.
It is not as of this commit, therefore the implementation is left TODO.

In addition this worker-spanwing code is added to AppSec only as AppSec
is currently the only user of remote configuration. This may be change
in the future, e.g by adding a similar barrier to Tracing, or spawning
the worker in other ways. The barrier should remain though as AppSec is
anticipated to e.g be able to work without Tracing.

In all cases this declarative barrier is a forward-compatible,
encapsulating API for which the underpinnings can be made to evolve.