Glue remote configuration with AppSec processor lifecycle
This is implemented with an AppSec-specific Receiver block. This
Receiver as well as other parts such as features and capabilites will
factored out of Core and into AppSec in a later PR.

Reconfiguration is done internally to the AppSec component: there is no
AppSec component rebuilding, and configuration settings are untouched.

During reconfiguration, a Mutex protects Processor swapping and
finalizing.

This Mutex is exposed and guarantees the getting and usage of a valid
processor to request a new context. Once the context is created,
libddwaf internally keeps a reference to the handle to guarantee a
context is usable beyond handle destroying (which is triggered by
finalization)

An `@app.call(env)` case is hoisted out of the lock to prevent the lock
being held for the duration of the request should the processor not be
ready.