Add a basic `shell.nix` file
This is entirely optional to use but for nixpkgs users this is a really
easy way to get things up and running. With just:

```
nix-shell
```

All needed dependencies get automatically pulled in, dropping the user
in a ready-to-use shell.

Then as usual:

```
bundle install
```

Puts everything in a `vendor/bundle` directory, which behaves as usual
regarding `gem install` and `bundle` in general. Should it need to be
cleaned up, e.g to start form scratch, a simple `rm -rf vendor/bundle`
is sufficient.

Also, gem commands can typically be run with or without `bundle exec`.

It also pins to a specific commit for the channel, so that tool versions
are always consistent across platforms (incl. CPU arches and OSes).