Add workaround for pending gitlab benchmark jobs
**What does this PR do?**:

This PR adds a workaround for the GitLab benchmark jobs configuration
that was causing them to show up on GitHub as "Pending" and never
completing for every PR.

**Motivation**:

Currently the jobs don't run on every PR so this fixes the "Pending"
issue.

Hopefully soon we'll start running the benchmarks on every PR and
so we'll be able to remove this workaround.

**Additional Notes**:

Thanks @ddyurchenko for the help setting this so far.

**How to test the change?**:

Verify that you no longer see "Pending" jobs on the GitHub CI status.