[PROF-7361] Implement "no signals" workaround for running CPU Profiling 2.0 without signals
**What does this PR do?**:

This PR adds a new mode to the profiler, which we're calling the
"no signals" workaround.

**Motivation**:

This mode is a workaround for a few incompatibility issues with the new
CPU Profiling 2.0 profiler. In #2702, when we made CPU Profiling 2.0
the default mode, we listed two cases where we were falling back to the
legacy profiler:

* [Some gems use native libraries that don't handle the `EINTR` error
  code that can be caused by a signal handler interrupting a system
  call](https://docs.datadoghq.com/profiler/profiler_troubleshooting/ruby/#unexpected-run-time-failures-and-errors-from-ruby-gems-that-use-native-extensions-in-dd-trace-rb-1110)
* Ruby 2.5 or below are missing an API that allows the profiler to
  detect if the current thread is holding on to the Global VM Lock
  in the signal handler, which we suspect may lead to very rare crashes

The objective of the "no signals" workaround is to allow the new
CPU Profiling 2.0 to be used by customers in the above situations,
both of which are related to signal handling.

Thus, we're able to migrate every customer to the new codebase,
allowing them to access the new features provided by the profiler,
and allowing us to completely retire the legacy profiler.

(I'll still keep the legacy profiler alive for a couple more releases,
but the plan is to delete it soon).

**Additional Notes**:

The "no signals workaround" mode works by making the CPU Profiling 2.0
profiler behave in a similar way to the legacy profiler when deciding
to take samples.

That is, the CPU Profiling 2.0 profiler by default runs in a
background thread that **does not hold the global VM lock** and thus
is running in parallel with the Ruby application.

When it decides it wants to take a sample, it sends a unix SIGPROF
signal to the currently-active Ruby thread, temporarily interrupting what
that thread is doing to ask the VM to take a sample.

This allows us to get higher quality data, because it's the profiler
that decides when it wants to take a sample and interrupts the code
to do so.

The "no signals workaround" instead modifies the above to have the
profiler **grab the global VM lock** when it wants to take a sample,
instead of interrupting the currently-active Ruby thread.

This means that instead of sampling immediately, the profiler needs
to wait for a Ruby VM scheduling switch point, and for the Ruby VM
scheduler to actually schedule our thread to execute.

This means that the profiler may need to wait dozens or hundreds
of milliseconds to sample. But... it works!

And the above is effectively how the legacy profiler worked -- since
it was written in pure Ruby code, it always needed to wait for its
turn to run.

Implementing the new behavior is not a lot of work because we already
had the scaffolding to do it as part of being able to sample in
periods of idleness (see #2468). So most of the work on this was
validating and adding the needed configuration.

**How to test the change?**:

Validate that in the situations described above where the legacy
profiler would be used no longer happen -- instead, the new
profiler gets used, and the "no signals workaround" gets enabled.